requirements.txt
httpx
logfire
devtools
python-dotenv
pydantic-ai
google-generativeai
gradio>=5.9.0
nest-asyncio
aioconsole